server/requirement.txt
anyio==3.6.2
asgiref==3.5.2
click==8.1.3
Django==4.1.1
fastapi==0.95.0
h11==0.14.0
httptools==0.5.0
idna==3.4
mysql-connector-python==8.0.32
protobuf==3.20.3
pydantic==1.10.7
pypng==0.20220715.0
python-dotenv==1.0.0
PyYAML==6.0
qrcode==7.4.2
sniffio==1.3.0
SQLAlchemy==2.0.7
sqlparse==0.4.3
starlette==0.26.1
typing_extensions==4.5.0
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
